Fix: 코드리뷰 반영 수정

+++ src/main/java/com/team3/deokhugam/service/s3/S3Service.java
@@ -1,10 +1,13 @@
 package com.team3.deokhugam.service.s3;
 
+import com.team3.deokhugam.exception.s3.EmptyFileUploadException;
+import com.team3.deokhugam.exception.s3.InvalidRequestException;
 import com.team3.deokhugam.exception.s3.S3DeleteException;
 import com.team3.deokhugam.exception.s3.S3UploadException;
 import com.team3.deokhugam.global.config.AwsProperties;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.core.sync.RequestBody;
@@ -15,6 +18,7 @@ import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.S3Exception;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 public class S3Service {
 
@@ -22,6 +26,12 @@ public class S3Service {
   private final AwsProperties props;
 
   public String upload(MultipartFile file, String key) {
+    log.info("파일 업로드 시작 - 요청 KEY: {}", key);
+    if (file == null || file.isEmpty()) {
+      log.debug("빈 파일 업로드 시도 발생");
+      throw new EmptyFileUploadException();
+    }
+
     try {
       s3Client.putObject(
           PutObjectRequest.builder()
@@ -33,11 +43,19 @@ public class S3Service {
       );
       return generateUrl(key);
     } catch (S3Exception | IOException e) {
+      log.error("S3 파일 업로드 실패 - KEY: {}", key, e);
       throw new S3UploadException();
     }
   }
 
   public void delete(String key) {
+    log.info("S3 파일 삭제 시작 - 요청 KEY: {}", key);
+
+    if (key == null || key.isBlank()) {
+      log.info("유효하지 않은 S3 삭제 키 전달됨");
+      throw new InvalidRequestException();
+    }
+
     try {
       s3Client.deleteObject(
           DeleteObjectRequest.builder()
@@ -45,7 +63,9 @@ public class S3Service {
               .key(key)
               .build()
       );
+      log.info("S3 파일 삭제 완료 - KEY: {}", key);
     } catch (S3Exception e) {
+      log.error("S3 파일 삭제 실패 - KEY: {}", key, e);
       throw new S3DeleteException();
     }
   }
